feat: add ocr endpoint functionality
Added comicTextDetector files.

Included an assets folder with subdirectores for each step of the OCR process.

Implemented a text detection routine to obtain text areas coordinates, as well as other routines for clipping the text from the image, scanning the text from the clips with an OCR library and clearing the files created during the process.